user events refactor to wallet events
